requirements.txt
asttokens==2.4.1
bcrypt==3.2.0
cffi==1.17.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
ffmpeg-python==0.2.0
future==1.0.0
importlib-metadata==8.2.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter-client==8.6.2
jupyter-core==5.7.2
matplotlib-inline==0.1.7
natsort==8.4.0
nest-asyncio==1.6.0
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.3
pvrecorder==1.2.2
PyAudio==0.2.14
pycparser==2.22
pygments==2.18.0
pyscreenrec==0.4
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.1.0
six==1.16.0
SoundCard==0.4.3
soundfile==0.12.1
stack-data==0.6.3
tk==0.1.0
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
zipp==3.20.0
